ASoC: wm_adsp: init host buffer struct when fw is loaded
If the firmware has a host transfer buffer, initialize our
view of it when the firmware is loaded instead of waiting
until a data stream is opened. This allows us to check the
firmware status without needing a compressed stream open.

The host_buffer_info is invalidated if the DSP is powered down,
since it's possible to power-down the DSP while the compressed
stream is still active (for example if the audio input is
disconnected.)

Change-Id: I0f8821f046b354bd510f8b72e091f542f1d28016
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>